feat: remove reflectable from app
