+++ packages/flutter_nekoton_bridge/example/README.md
@@ -2,13 +2,3 @@
 
 Demonstrates how to use the flutter_nekoton_bridge plugin.
 
-## Getting Started
-
-To generate reflectable shings run:
-
-```
-dart run build_runner build
-```
-
-TODO: Integrate this thing into melos and github actions.
-
