+++ packages/flutter_nekoton_bridge/example/pubspec.yaml
@@ -44,8 +44,6 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   freezed_annotation: ^2.2.0
-  reflectable: ^4.0.5
-
 
 dev_dependencies:
   flutter_test:
