+++ packages/flutter_nekoton_bridge/pubspec_overrides.yaml
@@ -1,5 +1,8 @@
+# dependency_overrides:
+#   nekoton_bridge:
+#     git:
+#       url: https://github.com/broxus/nekoton_bridge.git
+#       path: ../packages/nekoton_bridge
 dependency_overrides:
   nekoton_bridge:
-    git:
-      url: https://github.com/broxus/nekoton_bridge.git
-      path: ../packages/nekoton_bridge
+    path: ../nekoton_bridge
